// lib/earth3d/proto/rocktree.ts
export interface INodeKey {
    path?: string;
    epoch?: number;
}
// ...
export interface INodeData {
    matrix_globe_from_mesh?: number[];
    meshes?: IMesh[];
    copyright_ids?: number[];
    node_key?: INodeKey;
    kml_bounding_box?: number[];
    water_mesh?: IMesh;
    overlay_surface_meshes?: IMesh[];
    for_normals?: Uint8Array;
}

export interface IMesh {
    vertices?: Uint8Array;
    vertex_alphas?: Uint8Array;
    texture_coords?: Uint8Array;
    indices?: Uint8Array;
    octant_ranges?: Uint8Array;
    layer_counts?: Uint8Array;
    texture?: ITexture[];
    texture_coordinates?: Uint8Array;
    uv_offset_and_scale?: number[];
    layer_and_octant_counts?: Uint8Array;
    normals?: Uint8Array;
    normals_dev?: Uint8Array;
    mesh_id?: number;
    skirt_flags?: Uint8Array;
}
// ...
export interface ITexture {
    data?: Uint8Array[];
    format?: Texture_Format_Value;
    width?: number;
    height?: number;
    view_direction?: Texture_ViewDirection_Value;
    mesh_id?: number;
}